Moved All acceptance tests to IO.Ably.Specs and removed the other test project
