--- src/Ably.AcceptanceTests/Key.cs
@@ -1,12 +1,0 @@
-using System;
-
-namespace IO.Ably.AcceptanceTests
-{
-    public class Key
-    {
-        public String keyName;
-        public String keySecret;
-        public String keyStr;
-        public String capability;
-    }
-}
