--- src/Ably.AcceptanceTests/StatsFixture.json
@@ -1,22 +1,0 @@
-﻿[
-    {
-      "intervalId": "[[Interval1]]",
-      "inbound":  { "realtime": { "messages": { "count": 50, "data": 5000 } } },
-      "outbound": { "realtime": { "messages": { "count": 20, "data": 2000 } } }
-    },
-    {
-      "intervalId": "[[Interval2]]",
-      "inbound":  { "realtime": { "messages": { "count": 60, "data": 6000 } } },
-      "outbound": { "realtime": { "messages": { "count": 10, "data": 1000 } } }
-    },
-    {
-      "intervalId": "[[Interval3]]",
-      "inbound":       { "realtime": { "messages": { "count": 70, "data": 7000 } } },
-      "outbound":      { "realtime": { "messages": { "count": 40, "data": 4000 } } },
-      "persisted":     { "presence": { "count": 20, "data": 2000 } },
-      "connections":   { "tls":      { "peak": 20,  "opened": 10 } },
-      "channels":      { "peak": 50, "opened": 30 },
-      "apiRequests":   { "succeeded": 50, "failed": 10 },
-      "tokenRequests": { "succeeded": 60, "failed": 20 }
-    }
-  ]
