--- src/Ably.AcceptanceTests/crypto-data-128.json
@@ -1,57 +1,0 @@
-﻿
-{
-  "algorithm": "aes",
-  "mode": "cbc",
-  "keylength": 128,
-  "key": "WUP6u0K7MXI5Zeo0VppPwg==",
-  "iv": "HO4cYSP8LybPYBPZPHQOtg==",
-  "items": [
-    {
-      "encoded": {
-        "name": "example",
-        "data": "The quick brown fox jumped over the lazy dog"
-      },
-      "encrypted": {
-        "name": "example",
-        "data": "HO4cYSP8LybPYBPZPHQOtmHItcxYdSvcNUC6kXVpMn0VFL+9z2/5tJ6WFbR0SBT1xhFRuJ+MeBGTU3yOY9P5ow==",
-        "encoding": "utf-8/cipher+aes-128-cbc/base64"
-      }
-    },
-    {
-      "encoded": {
-        "name": "example",
-        "data": "AAECAwQFBgcICQoLDA0ODw==",
-        "encoding": "base64"
-      },
-      "encrypted": {
-        "name": "example",
-        "data": "HO4cYSP8LybPYBPZPHQOtuB3dfKG08yw7J4qx3kkjxdW0eoZv+nGAp76OKqYQ327",
-        "encoding": "cipher+aes-128-cbc/base64"
-      }
-    },
-    {
-      "encoded": {
-        "name": "example",
-        "data": "{\"example\":{\"json\":\"Object\"}}",
-        "encoding": "json"
-      },
-      "encrypted": {
-        "name": "example",
-        "data": "HO4cYSP8LybPYBPZPHQOtuD53yrD3YV3NBoTEYBh4U0N1QXHbtkfsDfTspKeLQFt",
-        "encoding": "json/utf-8/cipher+aes-128-cbc/base64"
-      }
-    },
-    {
-      "encoded": {
-        "name": "example",
-        "data": "[\"example\",\"json\",\"array\"]",
-        "encoding": "json"
-      },
-      "encrypted": {
-        "name": "example",
-        "data": "HO4cYSP8LybPYBPZPHQOtvmStzmExkdjvrn51J6cmaTZrGl+EsJ61sgxmZ6j6jcA",
-        "encoding": "json/utf-8/cipher+aes-128-cbc/base64"
-      }
-    }
-  ]
-}
