--- src/Ably.AcceptanceTests/crypto-data-256.json
@@ -1,57 +1,0 @@
-﻿
-{
-  "algorithm": "aes",
-  "mode": "cbc",
-  "keylength": 256,
-  "key": "o9qXZoPGDNla50VnRwH7cGqIrpyagTxGsRgimKJbY40=",
-  "iv": "NMDl1Acnel8HVdu1cEWdrw==",
-  "items": [
-    {
-      "encoded": {
-        "name": "example",
-        "data": "The quick brown fox jumped over the lazy dog"
-      },
-      "encrypted": {
-        "name": "example",
-        "data": "NMDl1Acnel8HVdu1cEWdr9CGPYFoBoLgJCzoybbQbnyfwx3UQ8CGuKyP/g56Za/JB3xW6XGkNzrHYvZwad4fvA==",
-        "encoding": "utf-8/cipher+aes-256-cbc/base64"
-      }
-    },
-    {
-      "encoded": {
-        "name": "example",
-        "data": "AAECAwQFBgcICQoLDA0ODw==",
-        "encoding": "base64"
-      },
-      "encrypted": {
-        "name": "example",
-        "data": "NMDl1Acnel8HVdu1cEWdr8UFEi56Ms0zPHszbppM61BC8Yf6ndq+kiCj9xXW97/O",
-        "encoding": "cipher+aes-256-cbc/base64"
-      }
-    },
-    {
-      "encoded": {
-        "name": "example",
-        "data": "{\"example\":{\"json\":\"Object\"}}",
-        "encoding": "json"
-      },
-      "encrypted": {
-        "name": "example",
-        "data": "NMDl1Acnel8HVdu1cEWdr21pS5//hdtQf3QqQzZM/jWAtn09Vh52E6jMdC3mWS98",
-        "encoding": "json/utf-8/cipher+aes-256-cbc/base64"
-      }
-    },
-    {
-      "encoded": {
-        "name": "example",
-        "data": "[\"example\",\"json\",\"array\"]",
-        "encoding": "json"
-      },
-      "encrypted": {
-        "name": "example",
-        "data": "NMDl1Acnel8HVdu1cEWdr4J5sVAFpnXsz0fTtsuwOaTRU+6P5GaWlNjePWwiOZCQ",
-        "encoding": "json/utf-8/cipher+aes-256-cbc/base64"
-      }
-    }
-  ]
-}
