--- src/Ably.AcceptanceTests/packages.config
@@ -1,7 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="FluentAssertions" version="4.3.2" targetFramework="net45" />
-  <package id="MsgPack.Cli" version="0.6.8" targetFramework="net45" />
-  <package id="Newtonsoft.Json" version="8.0.3" targetFramework="net45" />
-  <package id="NUnit" version="3.2.0" targetFramework="net45" />
-</packages>
