--- src/Ably.AcceptanceTests/testAppSpec.json
@@ -1,22 +1,0 @@
-{
-    "keys": [
-        { }
-    ],
-    "namespaces": [
-        {
-            "id": "persisted",
-            "persisted": true
-        }
-    ],
-    "channels": [
-        {
-            "name": "persisted:presence_fixtures",
-            "presence": [
-                { "clientId": "client_bool", "data": "true" },
-                { "clientId": "client_int", "data": "24" },
-                { "clientId": "client_string", "data": "This is a string clientData payload" },
-                { "clientId": "client_json", "data": "{ \"test\" => \"This is a JSONObject clientData payload\"}" }
-            ]
-        }
-    ]
-}
