+++ src/IO.Ably.Tests/Rest/StatsSpecs.cs
@@ -1,11 +1,134 @@
+using System;
+using System.Linq;
 using System.Net.Http;
 using System.Threading.Tasks;
+using FluentAssertions;
 using Xunit;
 using Xunit.Abstractions;
 using static Xunit.Assert;
 
 namespace IO.Ably.Tests
 {
+    [Collection("AblyRest SandBox Collection")]
+    [Trait("requires", "sandbox")]
+    public class StatsSandBoxSpecs : SandboxSpecs
+    {
+        public readonly static DateTimeOffset StartInterval = DateHelper.CreateDate(DateTimeOffset.UtcNow.Year - 1, 2, 3, 15, 5);
+
+        //Stats fixtures can be found in StatsFixture.json which is posted to /stats in TestsSetup.cs
+
+        private readonly Protocol _protocol;
+
+        private AblyRest GetAbly()
+        {
+            var testData = TestsSetup.TestData;
+
+            var options = new ClientOptions
+            {
+                Key = testData.keys.First().keyStr,
+                UseBinaryProtocol = _protocol == Protocol.MsgPack,
+                Environment = AblyEnvironment.Sandbox
+            };
+            var ably = new AblyRest(options);
+            return ably;
+        }
+
+        [TestFixture(Protocol.Json)]
+        [TestFixture(Protocol.MsgPack)]
+        //[Ignore("Ignoring for the time being. Will fix when I get to stats")]
+        public class ByMinuteWhenFromSetToStartIntervalAndLimitSetTo1 : StatsAcceptanceTests
+        {
+            public ByMinuteWhenFromSetToStartIntervalAndLimitSetTo1(Protocol protocol) : base(protocol)
+            {
+            }
+
+            [OneTimeSetUp]
+            public void GetStats()
+            {
+                var client = GetAbly();
+                Stats = client.Stats(new StatsDataRequestQuery() { Start = StartInterval.AddMinutes(-30), Limit = 1 }).Result;
+
+                TestStats = Stats.First();
+            }
+
+            public PaginatedResult<Stats> Stats { get; set; }
+            public Stats TestStats { get; set; }
+
+            [Fact]
+            public void RetrievesOnlyOneStat()
+            {
+                Stats.Should().HaveCount(1);
+            }
+
+            [Fact]
+            public void ShouldReturnCorrectAggregatedMessageData()
+            {
+                TestStats.All.Messages.Count.Should().Be(40 + 70);
+                TestStats.All.Messages.Data.Should().Be(4000 + 7000);
+            }
+
+            [Fact]
+            public void ReturnsAccurateInboundRealtimeAllData()
+            {
+                TestStats.Inbound.Realtime.All.Count.Should().Be(70);
+                TestStats.Inbound.Realtime.All.Data.Should().Be(7000);
+            }
+
+            [Fact]
+            public void ReturnsAccurateInboundRealtimeMessageData()
+            {
+                TestStats.Inbound.Realtime.Messages.Count.Should().Be(70);
+                TestStats.Inbound.Realtime.Messages.Data.Should().Be(7000);
+            }
+
+            [Fact]
+            public void ReturnsAccurateOutboundRealtimeAllData()
+            {
+                TestStats.Outbound.Realtime.All.Count.Should().Be(40);
+                TestStats.Outbound.Realtime.All.Data.Should().Be(4000);
+            }
+
+            [Fact]
+            public void ReturnsAccuratePersistedPresenceAllData()
+            {
+                TestStats.Persisted.Presence.Count.Should().Be(20);
+                TestStats.Persisted.Presence.Data.Should().Be(2000);
+            }
+
+            [Fact]f
+            public void ReturnsAccurateConnectionsAllData()
+            {
+                TestStats.Connections.Tls.Peak.Should().Be(20);
+                TestStats.Connections.Tls.Opened.Should().Be(10);
+            }
+
+            [Fact]
+            public void ReturnsAccurateChannelsAllData()
+            {
+                TestStats.Channels.Peak.Should().Be(50);
+                TestStats.Channels.Opened.Should().Be(30);
+            }
+
+            [Fact]
+            public void ReturnsAccurentApiRequestsData()
+            {
+                TestStats.ApiRequests.Succeeded.Should().Be(50);
+                TestStats.ApiRequests.Failed.Should().Be(10);
+            }
+
+            [Fact]
+            public void ReturnsAccurateTokenRequestsData()
+            {
+                TestStats.TokenRequests.Succeeded.Should().Be(60);
+                TestStats.TokenRequests.Failed.Should().Be(20);
+            }
+        }
+
+        public StatsSandBoxSpecs(AblySandboxFixture fixture, ITestOutputHelper output) : base(fixture, output)
+        {
+        }
+    }
+
     public class StatsSpecs : MockHttpRestSpecs
     {
         [Fact]
